Backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
asyncpg
pgvector
pydantic
pydantic[email]
python-jose[cryptography]
passlib[argon2]
python-multipart
numpy
playwright
selectolax
sentence-transformers
torch
aiosmtplib
httpx
scalar-fastapi
